+++ hoomd/hpmc/GSDHPMCSchema.h
@@ -1,3 +1,5 @@
+// Copyright (c) 2009-2022 The Regents of the University of Michigan.
+// Part of HOOMD-blue, released under the BSD 3-Clause License.
 
 #include "ShapeConvexPolygon.h"    // check
 #include "ShapeConvexPolyhedron.h" // check