+++ hoomd/hpmc/module_union_convex_polyhedron.cc
@@ -1,5 +1,5 @@
-// Copyright (c) 2009-2021 The Regents of the University of Michigan
-// This file is part of the HOOMD-blue project, released under the BSD 3-Clause License.
+// Copyright (c) 2009-2022 The Regents of the University of Michigan.
+// Part of HOOMD-blue, released under the BSD 3-Clause License.
 
 // Include the defined classes that are to be exported to python
 #include "ComputeFreeVolume.h"
@@ -12,7 +12,7 @@
 
 #include "ExternalField.h"
 #include "ExternalFieldComposite.h"
-#include "ExternalFieldLattice.h"
+#include "ExternalFieldHarmonic.h"
 #include "ExternalFieldWall.h"
 
 #include "UpdaterClusters.h"
@@ -51,9 +51,9 @@ void export_union_convex_polyhedron(pybind11::module& m)
     export_ExternalFieldInterface<ShapeUnion<ShapeSpheropolyhedron>>(
         m,
         "ExternalFieldConvexPolyhedronUnion");
-    export_LatticeField<ShapeUnion<ShapeSpheropolyhedron>>(
+    export_HarmonicField<ShapeUnion<ShapeSpheropolyhedron>>(
         m,
-        "ExternalFieldLatticeConvexPolyhedronUnion");
+        "ExternalFieldHarmonicConvexPolyhedronUnion");
     export_ExternalFieldComposite<ShapeUnion<ShapeSpheropolyhedron>>(
         m,
         "ExternalFieldCompositeConvexPolyhedronUnion");