+++ hoomd/md/WallData.h
@@ -1,14 +1,12 @@
-// Copyright (c) 2009-2021 The Regents of the University of Michigan
-// This file is part of the HOOMD-blue project, released under the BSD 3-Clause License.
-
-// Maintainer: jproc
+// Copyright (c) 2009-2022 The Regents of the University of Michigan.
+// Part of HOOMD-blue, released under the BSD 3-Clause License.
 
 /*! \file WallData.h
     \brief Contains declarations for all types (currently Sphere, Cylinder, and
     Plane) of WallData and associated utilities.
  */
-#ifndef WALL_DATA_H
-#define WALL_DATA_H
+
+#pragma once
 
 #include "hoomd/HOOMDMath.h"
 #include "hoomd/VectorMath.h"
@@ -17,6 +15,10 @@
 #define DEVICE __device__
 #else
 #define DEVICE
+#endif
+
+#ifndef __HIPCC__
+#include <pybind11/pybind11.h>
 #endif
 
 //! SphereWall Constructor
@@ -37,13 +39,17 @@ namespace md
     {
 struct __attribute__((visibility("default"))) SphereWall
     {
-    SphereWall(Scalar rad = 0.0, Scalar3 orig = make_scalar3(0.0, 0.0, 0.0), bool ins = true)
-        : origin(vec3<Scalar>(orig)), r(rad), inside(ins)
+    SphereWall(Scalar rad = 0.0,
+               Scalar3 orig = make_scalar3(0.0, 0.0, 0.0),
+               bool ins = true,
+               bool open_ = true)
+        : origin(vec3<Scalar>(orig)), r(rad), inside(ins), open(open_)
         {
         }
     vec3<Scalar> origin; // need to order datatype in descending order of type size for Fermi
     Scalar r;
     bool inside;
+    bool open;
     } __attribute__((aligned(ALIGN_SCALAR))); // align according to first member of vec3<Scalar>
 
 //! CylinderWall Constructor
@@ -59,8 +65,9 @@ struct __attribute__((visibility("default"))) CylinderWall
     CylinderWall(Scalar rad = 0.0,
                  Scalar3 orig = make_scalar3(0.0, 0.0, 0.0),
                  Scalar3 zorient = make_scalar3(0.0, 0.0, 1.0),
-                 bool ins = true)
-        : origin(vec3<Scalar>(orig)), axis(vec3<Scalar>(zorient)), r(rad), inside(ins)
+                 bool ins = true,
+                 bool open_ = true)
+        : origin(vec3<Scalar>(orig)), axis(vec3<Scalar>(zorient)), r(rad), inside(ins), open(open_)
         {
         vec3<Scalar> zVec = axis;
         vec3<Scalar> zNorm(0.0, 0.0, 1.0);
@@ -91,19 +98,19 @@ struct __attribute__((visibility("default"))) CylinderWall
     vec3<Scalar> axis;
     Scalar r;
     bool inside;
+    bool open;
     } __attribute__((aligned(ALIGN_SCALAR))); // align according to first member of quat<Scalar>
 
 //! PlaneWall Constructor
 /*! \param origin The x,y,z coordinates of a point on the cylinder axis
     \param normal The x,y,z normal vector of the plane (normalized upon input)
-    \param inside Determines which half space is evaluated.
 */
 struct __attribute__((visibility("default"))) PlaneWall
     {
     PlaneWall(Scalar3 orig = make_scalar3(0.0, 0.0, 0.0),
               Scalar3 norm = make_scalar3(0.0, 0.0, 1.0),
-              bool ins = true)
-        : normal(vec3<Scalar>(norm)), origin(vec3<Scalar>(orig)), inside(ins)
+              bool open_ = true)
+        : normal(vec3<Scalar>(norm)), origin(vec3<Scalar>(orig)), open(open_)
         {
         vec3<Scalar> nVec;
         nVec = normal;
@@ -113,70 +120,60 @@ struct __attribute__((visibility("default"))) PlaneWall
         }
     vec3<Scalar> normal;
     vec3<Scalar> origin;
-    bool inside;
+    bool open;
     } __attribute__((aligned(ALIGN_SCALAR))); // align according to first member of vec3<Scalar>
 
-//! Point to wall vector for a sphere wall geometry
+//! Wall vector to point for a sphere wall geometry
 /* Returns 0 vector when all normal directions are equal
  */
-DEVICE inline vec3<Scalar>
-vecPtToWall(const SphereWall& wall, const vec3<Scalar>& position, bool& inside)
+DEVICE inline Scalar3
+distVectorWallToPoint(const SphereWall& wall, const vec3<Scalar>& position, bool& in_active_space)
     {
-    vec3<Scalar> t = position;
-    t -= wall.origin;
-    vec3<Scalar> shiftedPos(t);
-    Scalar rxyz = sqrt(dot(shiftedPos, shiftedPos));
-    if (rxyz > 0.0)
+    const vec3<Scalar> dist_from_origin = position - wall.origin;
+    const Scalar euclidean_dist = sqrt(dot(dist_from_origin, dist_from_origin));
+    if (euclidean_dist == 0.0)
         {
-        inside = (((rxyz <= wall.r) && wall.inside) || ((rxyz > wall.r) && !(wall.inside))) ? true
-                                                                                            : false;
-        t *= wall.r / rxyz;
-        vec3<Scalar> dx = t - shiftedPos;
-        return dx;
+        in_active_space = wall.open;
+        return make_scalar3(0.0, 0.0, 0.0);
         }
-    else
-        {
-        inside = (wall.inside) ? true : false;
-        return vec3<Scalar>(0.0, 0.0, 0.0);
-        }
+    in_active_space
+        = ((euclidean_dist < wall.r) && wall.inside) || ((euclidean_dist > wall.r) && !wall.inside);
+    return vec_to_scalar3((1 - (wall.r / euclidean_dist)) * dist_from_origin);
     };
 
-//! Point to wall vector for a cylinder wall geometry
+//! Wall vector to point for a cylinder wall geometry
 /* Returns 0 vector when all normal directions are equal
  */
-DEVICE inline vec3<Scalar>
-vecPtToWall(const CylinderWall& wall, const vec3<Scalar>& position, bool& inside)
+DEVICE inline Scalar3
+distVectorWallToPoint(const CylinderWall& wall, const vec3<Scalar>& position, bool& in_active_space)
     {
-    vec3<Scalar> t = position;
-    t -= wall.origin;
-    vec3<Scalar> shiftedPos = rotate(wall.quatAxisToZRot, t);
-    shiftedPos.z = 0.0;
-    Scalar rxy = sqrt(dot(shiftedPos, shiftedPos));
-    if (rxy > 0.0)
+    const vec3<Scalar> dist_from_origin = position - wall.origin;
+    vec3<Scalar> rotated_distance = rotate(wall.quatAxisToZRot, dist_from_origin);
+    rotated_distance.z = 0.0;
+    const Scalar euclidean_dist = sqrt(dot(rotated_distance, rotated_distance));
+    if (euclidean_dist == 0.0)
         {
-        inside = (((rxy <= wall.r) && wall.inside) || ((rxy > wall.r) && !(wall.inside))) ? true
-                                                                                          : false;
-        t = (wall.r / rxy) * shiftedPos;
-        vec3<Scalar> dx = t - shiftedPos;
-        dx = rotate(conj(wall.quatAxisToZRot), dx);
-        return dx;
+        in_active_space = wall.open;
+        return make_scalar3(0.0, 0.0, 0.0);
         }
-    else
-        {
-        inside = (wall.inside) ? true : false;
-        return vec3<Scalar>(0.0, 0.0, 0.0);
-        }
+    in_active_space = ((euclidean_dist < wall.r) && wall.inside)
+                      || ((euclidean_dist > wall.r) && !(wall.inside));
+    const vec3<Scalar> dx = (1 - (wall.r / euclidean_dist)) * rotated_distance;
+    return vec_to_scalar3(rotate(conj(wall.quatAxisToZRot), dx));
     };
 
-//! Point to wall vector for a plane wall geometry
-DEVICE inline vec3<Scalar>
-vecPtToWall(const PlaneWall& wall, const vec3<Scalar>& position, bool& inside)
+//! Wall vector to point for a plane wall geometry
+DEVICE inline Scalar3
+distVectorWallToPoint(const PlaneWall& wall, const vec3<Scalar>& position, bool& in_active_space)
     {
-    vec3<Scalar> t = position;
-    Scalar d = dot(wall.normal, t) - dot(wall.normal, wall.origin);
-    inside = (((d >= 0.0) && wall.inside) || ((d < 0.0) && !(wall.inside))) ? true : false;
-    vec3<Scalar> dx = -d * wall.normal;
-    return dx;
+    Scalar distance = dot(wall.normal, position) - dot(wall.normal, wall.origin);
+    if (distance == 0)
+        {
+        in_active_space = wall.open;
+        return make_scalar3(0.0, 0.0, 0.0);
+        }
+    in_active_space = distance > 0.0;
+    return vec_to_scalar3(distance * wall.normal);
     };
 
 //! Distance of point to inside sphere wall geometry, not really distance, +- based on if it's
@@ -189,7 +186,7 @@ DEVICE inline Scalar distWall(const SphereWall& wall, const vec3<Scalar>& positi
     Scalar rxyz2
         = shiftedPos.x * shiftedPos.x + shiftedPos.y * shiftedPos.y + shiftedPos.z * shiftedPos.z;
     Scalar d = wall.r - sqrt(rxyz2);
-    d = (wall.inside) ? d : -d;
+    d = wall.inside ? d : -d;
     return d;
     };
 
@@ -202,7 +199,7 @@ DEVICE inline Scalar distWall(const CylinderWall& wall, const vec3<Scalar>& posi
     vec3<Scalar> shiftedPos = rotate(wall.quatAxisToZRot, t);
     Scalar rxy2 = shiftedPos.x * shiftedPos.x + shiftedPos.y * shiftedPos.y;
     Scalar d = wall.r - sqrt(rxy2);
-    d = (wall.inside) ? d : -d;
+    d = wall.inside ? d : -d;
     return d;
     };
 
@@ -210,11 +207,13 @@ DEVICE inline Scalar distWall(const CylinderWall& wall, const vec3<Scalar>& posi
 //! or not
 DEVICE inline Scalar distWall(const PlaneWall& wall, const vec3<Scalar>& position)
     {
-    vec3<Scalar> t = position;
-    Scalar d = dot(wall.normal, t) - dot(wall.normal, wall.origin);
-    d = (wall.inside) ? d : -d;
-    return d;
+    return dot(wall.normal, position) - dot(wall.normal, wall.origin);
     };
+
+#ifndef __HIPCC__
+// Export all wall data types into Python. This is needed to allow for syncing Python and C++
+// list/array data structures containing walls for WallPotential objects.
+void export_wall_data(pybind11::module& m);
+#endif
     } // end namespace md
     } // end namespace hoomd
-#endif