+++ hoomd/md/pair/__init__.py
@@ -1,6 +1,5 @@
-# Copyright (c) 2009-2021 The Regents of the University of Michigan
-# This file is part of the HOOMD-blue project, released under the BSD 3-Clause
-# License.
+# Copyright (c) 2009-2022 The Regents of the University of Michigan.
+# Part of HOOMD-blue, released under the BSD 3-Clause License.
 
 """Pair Potentials for molecular dynamics.
 
@@ -8,7 +7,7 @@ For anisotropic potentials see `hoomd.md.pair.aniso`
 """
 
 from . import aniso
-from .pair import (Pair, LJ, Gauss, SLJ, Yukawa, Ewald, Morse, DPD,
+from .pair import (Pair, LJ, Gauss, SLJ, ExpandedLJ, Yukawa, Ewald, Morse, DPD,
                    DPDConservative, DPDLJ, ForceShiftedLJ, Moliere, ZBL, Mie,
                    ExpandedMie, ReactionField, DLVO, Buckingham, LJ1208, LJ0804,
                    Fourier, OPP, Table, TWF)